Refactor Port and Connection Schemas + mapping helpers
