+++ equinix/data_source_fabric_port.go
@@ -10,7 +10,7 @@ import (
 func dataSourceFabricPort() *schema.Resource {
 	return &schema.Resource{
 		ReadContext: dataSourceFabricPortRead,
-		Schema:      readFabricPortResourceSchema(),
+		Schema:      FabricPortResourceSchema(),
 		Description: "Fabric V4 API compatible data resource that allow user to fetch port by uuid\n\n~> **Note** Equinix Fabric v4 resources and datasources are currently in Beta. The interfaces related to `equinix_fabric_` resources and datasources may change ahead of general availability. Please, do not hesitate to report any problems that you experience by opening a new [issue](https://github.com/equinix/terraform-provider-equinix/issues/new?template=bug.md)",
 	}
 }
