+++ equinix/fabric_mapping_helper.go
@@ -6,7 +6,6 @@ import (
 	"log"
 
 	v4 "github.com/equinix-labs/fabric-go/fabric/v4"
-	"github.com/equinix/terraform-provider-equinix/internal/converters"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -114,7 +113,7 @@ func accessPointToFabric(accessPointRequest []interface{}) v4.AccessPoint {
 		}
 
 		if len(locationList) != 0 {
-			sl := locationToFabric(locationList)
+			sl := equinix_schema.LocationToFabric(locationList)
 			accessPoint.Location = &sl
 		}
 
@@ -145,71 +144,6 @@ func cloudRouterToFabric(cloudRouterRequest []interface{}) v4.CloudRouter {
 	return cloudRouterMapped
 }
 
-func projectToFabric(projectRequest []interface{}) v4.Project {
-	if len(projectRequest) == 0 {
-		return v4.Project{}
-	}
-	mappedPr := v4.Project{}
-	for _, pr := range projectRequest {
-		prMap := pr.(map[string]interface{})
-		projectId := prMap["project_id"].(string)
-		mappedPr = v4.Project{ProjectId: projectId}
-	}
-	return mappedPr
-}
-
-func notificationToFabric(schemaNotifications []interface{}) []v4.SimplifiedNotification {
-	if schemaNotifications == nil {
-		return []v4.SimplifiedNotification{}
-	}
-	var notifications []v4.SimplifiedNotification
-	for _, n := range schemaNotifications {
-		ntype := n.(map[string]interface{})["type"].(string)
-		interval := n.(map[string]interface{})["send_interval"].(string)
-		emailsRaw := n.(map[string]interface{})["emails"].([]interface{})
-		emails := converters.IfArrToStringArr(emailsRaw)
-		notifications = append(notifications, v4.SimplifiedNotification{
-			Type_:        ntype,
-			SendInterval: interval,
-			Emails:       emails,
-		})
-	}
-	return notifications
-}
-
-func redundancyToFabric(schemaRedundancy []interface{}) v4.ConnectionRedundancy {
-	if schemaRedundancy == nil {
-		return v4.ConnectionRedundancy{}
-	}
-	red := v4.ConnectionRedundancy{}
-	for _, r := range schemaRedundancy {
-		redundancyMap := r.(map[string]interface{})
-		connectionPriority := v4.ConnectionPriority(redundancyMap["priority"].(string))
-		redundancyGroup := redundancyMap["group"].(string)
-		red = v4.ConnectionRedundancy{
-			Priority: &connectionPriority,
-			Group:    redundancyGroup,
-		}
-	}
-	return red
-}
-
-func orderToFabric(schemaOrder []interface{}) v4.Order {
-	if schemaOrder == nil {
-		return v4.Order{}
-	}
-	order := v4.Order{}
-	for _, o := range schemaOrder {
-		orderMap := o.(map[string]interface{})
-		purchaseOrderNumber := orderMap["purchase_order_number"]
-		billingTier := orderMap["billing_tier"]
-		orderId := orderMap["order_id"]
-		orderNumber := orderMap["order_number"]
-		order = v4.Order{PurchaseOrderNumber: purchaseOrderNumber.(string), BillingTier: billingTier.(string), OrderId: orderId.(string), OrderNumber: orderNumber.(string)}
-	}
-	return order
-}
-
 func linkProtocolToFabric(linkProtocolList []interface{}) v4.SimplifiedLinkProtocol {
 	slp := v4.SimplifiedLinkProtocol{}
 	for _, lp := range linkProtocolList {
@@ -224,15 +158,6 @@ func linkProtocolToFabric(linkProtocolList []interface{}) v4.SimplifiedLinkProto
 	return slp
 }
 
-func portToFabric(portList []interface{}) v4.SimplifiedPort {
-	p := v4.SimplifiedPort{}
-	for _, pl := range portList {
-		plMap := pl.(map[string]interface{})
-		uuid := plMap["uuid"].(string)
-		p = v4.SimplifiedPort{Uuid: uuid}
-	}
-	return p
-}
 func networkToFabric(networkList []interface{}) v4.SimplifiedNetwork {
 	p := v4.SimplifiedNetwork{}
 	for _, pl := range networkList {
@@ -254,23 +179,6 @@ func simplifiedServiceProfileToFabric(profileList []interface{}) v4.SimplifiedSe
 
 	}
 	return ssp
-}
-
-func locationToFabric(locationList []interface{}) v4.SimplifiedLocation {
-	sl := v4.SimplifiedLocation{}
-	for _, ll := range locationList {
-		llMap := ll.(map[string]interface{})
-		metroName := llMap["metro_name"]
-		var metroNamestr string
-		if metroName != nil {
-			metroNamestr = metroName.(string)
-		}
-		region := llMap["region"].(string)
-		mc := llMap["metro_code"].(string)
-		ibx := llMap["ibx"].(string)
-		sl = v4.SimplifiedLocation{MetroCode: mc, Region: region, Ibx: ibx, MetroName: metroNamestr}
-	}
-	return sl
 }
 
 func virtualdeviceToFabric(virtualdeviceList []interface{}) v4.VirtualDevice {
@@ -308,16 +216,6 @@ func accountToCloudRouter(accountList []interface{}) v4.SimplifiedAccount {
 	return sa
 }
 
-func locationToCloudRouter(locationList []interface{}) v4.SimplifiedLocationWithoutIbx {
-	sl := v4.SimplifiedLocationWithoutIbx{}
-	for _, ll := range locationList {
-		llMap := ll.(map[string]interface{})
-		mc := llMap["metro_code"].(string)
-		sl = v4.SimplifiedLocationWithoutIbx{MetroCode: mc}
-	}
-	return sl
-}
-
 func packageToCloudRouter(packageList []interface{}) v4.CloudRouterPackageType {
 	p := v4.CloudRouterPackageType{}
 	for _, pl := range packageList {
@@ -326,19 +224,6 @@ func packageToCloudRouter(packageList []interface{}) v4.CloudRouterPackageType {
 		p = v4.CloudRouterPackageType{Code: code}
 	}
 	return p
-}
-
-func projectToCloudRouter(projectRequest []interface{}) v4.Project {
-	if projectRequest == nil {
-		return v4.Project{}
-	}
-	mappedPr := v4.Project{}
-	for _, pr := range projectRequest {
-		prMap := pr.(map[string]interface{})
-		projectId := prMap["project_id"].(string)
-		mappedPr = v4.Project{ProjectId: projectId}
-	}
-	return mappedPr
 }
 
 func accountToTerra(account *v4.SimplifiedAccount) *schema.Set {
@@ -386,38 +271,6 @@ func accountCloudRouterToTerra(account *v4.SimplifiedAccount) *schema.Set {
 	return accountSet
 }
 
-func errorToTerra(errors []v4.ModelError) []interface{} {
-	if errors == nil {
-		return nil
-	}
-	mappedErrors := make([]interface{}, len(errors))
-	for index, mError := range errors {
-		mappedErrors[index] = map[string]interface{}{
-			"error_code":      mError.ErrorCode,
-			"error_message":   mError.ErrorMessage,
-			"correlation_id":  mError.CorrelationId,
-			"details":         mError.Details,
-			"help":            mError.Help,
-			"additional_info": errorAdditionalInfoToTerra(mError.AdditionalInfo),
-		}
-	}
-	return mappedErrors
-}
-
-func errorAdditionalInfoToTerra(additionalInfol []v4.PriceErrorAdditionalInfo) []interface{} {
-	if additionalInfol == nil {
-		return nil
-	}
-	mappedAdditionalInfol := make([]interface{}, len(additionalInfol))
-	for index, additionalInfo := range additionalInfol {
-		mappedAdditionalInfol[index] = map[string]interface{}{
-			"property": additionalInfo.Property,
-			"reason":   additionalInfo.Reason,
-		}
-	}
-	return mappedAdditionalInfol
-}
-
 func operationToTerra(operation *v4.ConnectionOperation) *schema.Set {
 	if operation == nil {
 		return nil
@@ -429,7 +282,7 @@ func operationToTerra(operation *v4.ConnectionOperation) *schema.Set {
 		mappedOperation["provider_status"] = string(*operation.ProviderStatus)
 		mappedOperation["equinix_status"] = string(*operation.EquinixStatus)
 		if operation.Errors != nil {
-			mappedOperation["errors"] = errorToTerra(operation.Errors)
+			mappedOperation["errors"] = equinix_schema.ErrorToTerra(operation.Errors)
 		}
 		mappedOperations = append(mappedOperations, mappedOperation)
 	}
@@ -438,139 +291,6 @@ func operationToTerra(operation *v4.ConnectionOperation) *schema.Set {
 		mappedOperations,
 	)
 	return operationSet
-}
-
-func orderMappingToTerra(order *v4.Order) *schema.Set {
-	if order == nil {
-		return nil
-	}
-	orders := []*v4.Order{order}
-	mappedOrders := make([]interface{}, len(orders))
-	for _, order := range orders {
-		mappedOrder := make(map[string]interface{})
-		mappedOrder["purchase_order_number"] = order.PurchaseOrderNumber
-		mappedOrder["billing_tier"] = order.BillingTier
-		mappedOrder["order_id"] = order.OrderId
-		mappedOrder["order_number"] = order.OrderNumber
-		mappedOrders = append(mappedOrders, mappedOrder)
-	}
-	orderSet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.OrderSch()}),
-		mappedOrders,
-	)
-	return orderSet
-}
-
-func changeLogToTerra(changeLog *v4.Changelog) *schema.Set {
-	if changeLog == nil {
-		return nil
-	}
-	changeLogs := []*v4.Changelog{changeLog}
-	mappedChangeLogs := make([]interface{}, len(changeLogs))
-	for _, changeLog := range changeLogs {
-		mappedChangeLog := make(map[string]interface{})
-		mappedChangeLog["created_by"] = changeLog.CreatedBy
-		mappedChangeLog["created_by_full_name"] = changeLog.CreatedByFullName
-		mappedChangeLog["created_by_email"] = changeLog.CreatedByEmail
-		mappedChangeLog["created_date_time"] = changeLog.CreatedDateTime.String()
-		mappedChangeLog["updated_by"] = changeLog.UpdatedBy
-		mappedChangeLog["updated_by_full_name"] = changeLog.UpdatedByFullName
-		mappedChangeLog["updated_date_time"] = changeLog.UpdatedDateTime.String()
-		mappedChangeLog["deleted_by"] = changeLog.DeletedBy
-		mappedChangeLog["deleted_by_full_name"] = changeLog.DeletedByFullName
-		mappedChangeLog["deleted_by_email"] = changeLog.DeletedByEmail
-		mappedChangeLog["deleted_date_time"] = changeLog.DeletedDateTime.String()
-		mappedChangeLogs = append(mappedChangeLogs, mappedChangeLog)
-	}
-	changeLogSet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.ChangeLogSch()}),
-		mappedChangeLogs,
-	)
-	return changeLogSet
-}
-
-func portRedundancyToTerra(redundancy *v4.PortRedundancy) *schema.Set {
-	redundancies := []*v4.PortRedundancy{redundancy}
-	mappedRedundancys := make([]interface{}, len(redundancies))
-	for _, redundancy := range redundancies {
-		mappedRedundancy := make(map[string]interface{})
-		mappedRedundancy["priority"] = string(*redundancy.Priority)
-		mappedRedundancys = append(mappedRedundancys, mappedRedundancy)
-	}
-	redundancySet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.RedundancySch()}),
-		mappedRedundancys,
-	)
-	return redundancySet
-}
-
-func redundancyToTerra(redundancy *v4.ConnectionRedundancy) *schema.Set {
-	if redundancy == nil {
-		return nil
-	}
-	redundancies := []*v4.ConnectionRedundancy{redundancy}
-	mappedRedundancys := make([]interface{}, len(redundancies))
-	for _, redundancy := range redundancies {
-		mappedRedundancy := make(map[string]interface{})
-		mappedRedundancy["group"] = redundancy.Group
-		mappedRedundancy["priority"] = string(*redundancy.Priority)
-		mappedRedundancys = append(mappedRedundancys, mappedRedundancy)
-	}
-	redundancySet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.RedundancySch()}),
-		mappedRedundancys,
-	)
-	return redundancySet
-}
-
-func notificationToTerra(notifications []v4.SimplifiedNotification) []map[string]interface{} {
-	if notifications == nil {
-		return nil
-	}
-	mappedNotifications := make([]map[string]interface{}, len(notifications))
-	for index, notification := range notifications {
-		mappedNotifications[index] = map[string]interface{}{
-			"type":          notification.Type_,
-			"send_interval": notification.SendInterval,
-			"emails":        notification.Emails,
-		}
-	}
-	return mappedNotifications
-}
-
-func locationToTerra(location *v4.SimplifiedLocation) *schema.Set {
-	locations := []*v4.SimplifiedLocation{location}
-	mappedLocations := make([]interface{}, len(locations))
-	for i, location := range locations {
-		mappedLocations[i] = map[string]interface{}{
-			"region":     location.Region,
-			"metro_name": location.MetroName,
-			"metro_code": location.MetroCode,
-			"ibx":        location.Ibx,
-		}
-	}
-	locationSet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.LocationSch()}),
-		mappedLocations,
-	)
-	return locationSet
-}
-
-func locationCloudRouterToTerra(location *v4.SimplifiedLocationWithoutIbx) *schema.Set {
-	locations := []*v4.SimplifiedLocationWithoutIbx{location}
-	mappedLocations := make([]interface{}, len(locations))
-	for i, location := range locations {
-		mappedLocations[i] = map[string]interface{}{
-			"region":     location.Region,
-			"metro_name": location.MetroName,
-			"metro_code": location.MetroCode,
-		}
-	}
-	locationSet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.LocationSch()}),
-		mappedLocations,
-	)
-	return locationSet
 }
 
 func serviceTokenToTerra(serviceToken *v4.ServiceToken) *schema.Set {
@@ -661,43 +381,6 @@ func cloudRouterPackageToTerra(packageType *v4.CloudRouterPackageType) *schema.S
 	return packageSet
 }
 
-func orderToTerra(order *v4.Order) *schema.Set {
-	if order == nil {
-		return nil
-	}
-	orders := []*v4.Order{order}
-	mappedOrders := make([]interface{}, len(orders))
-	for _, order := range orders {
-		mappedOrder := make(map[string]interface{})
-		mappedOrder["purchase_order_number"] = order.PurchaseOrderNumber
-		mappedOrder["billing_tier"] = order.BillingTier
-		mappedOrder["order_id"] = order.OrderId
-		mappedOrder["order_number"] = order.OrderNumber
-		mappedOrders = append(mappedOrders, mappedOrder)
-	}
-	orderSet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: equinix_schema.OrderSch()}),
-		mappedOrders)
-	return orderSet
-}
-
-func projectToTerra(project *v4.Project) *schema.Set {
-	if project == nil {
-		return nil
-	}
-	projects := []*v4.Project{project}
-	mappedProjects := make([]interface{}, len(projects))
-	for _, project := range projects {
-		mappedProject := make(map[string]interface{})
-		mappedProject["project_id"] = project.ProjectId
-		mappedProjects = append(mappedProjects, mappedProject)
-	}
-	projectSet := schema.NewSet(
-		schema.HashResource(readGatewayProjectSchRes),
-		mappedProjects)
-	return projectSet
-}
-
 func virtualDeviceToTerra(virtualDevice *v4.VirtualDevice) *schema.Set {
 	if virtualDevice == nil {
 		return nil
@@ -749,7 +432,7 @@ func accessPointToTerra(accessPoint *v4.AccessPoint) *schema.Set {
 			mappedAccessPoint["account"] = accountToTerra(accessPoint.Account)
 		}
 		if accessPoint.Location != nil {
-			mappedAccessPoint["location"] = locationToTerra(accessPoint.Location)
+			mappedAccessPoint["location"] = equinix_schema.LocationToTerra(accessPoint.Location)
 		}
 		if accessPoint.Port != nil {
 			mappedAccessPoint["port"] = portToTerra(accessPoint.Port)
@@ -885,26 +568,6 @@ func supportedBandwidthsToTerra(supportedBandwidths *[]int32) []interface{} {
 		mappedSupportedBandwidths = append(mappedSupportedBandwidths, int(bandwidth))
 	}
 	return mappedSupportedBandwidths
-}
-
-func portToTerra(port *v4.SimplifiedPort) *schema.Set {
-	ports := []*v4.SimplifiedPort{port}
-	mappedPorts := make([]interface{}, len(ports))
-	for _, port := range ports {
-		mappedPort := make(map[string]interface{})
-		mappedPort["href"] = port.Href
-		mappedPort["name"] = port.Name
-		mappedPort["uuid"] = port.Uuid
-		if port.Redundancy != nil {
-			mappedPort["redundancy"] = portRedundancyToTerra(port.Redundancy)
-		}
-		mappedPorts = append(mappedPorts, mappedPort)
-	}
-	portSet := schema.NewSet(
-		schema.HashResource(&schema.Resource{Schema: portSch()}),
-		mappedPorts,
-	)
-	return portSet
 }
 
 func routingProtocolDirectIpv4ToFabric(routingProtocolDirectIpv4Request []interface{}) v4.DirectConnectionIpv4 {
@@ -1140,7 +803,7 @@ func routingProtocolOperationToTerra(routingProtocolOperation *v4.RoutingProtoco
 	for _, routingProtocolOperation := range routingProtocolOperations {
 		mappedRpOperation := make(map[string]interface{})
 		if routingProtocolOperation.Errors != nil {
-			mappedRpOperation["errors"] = errorToTerra(routingProtocolOperation.Errors)
+			mappedRpOperation["errors"] = equinix_schema.ErrorToTerra(routingProtocolOperation.Errors)
 		}
 		mappedRpOperations = append(mappedRpOperations, mappedRpOperation)
 	}
