+++ equinix/fabric_routing_protocol_read_schema.go
@@ -190,7 +190,7 @@ func readRoutingProtocolOperationSch() map[string]*schema.Schema {
 			Computed:    true,
 			Description: "Errors occurred",
 			Elem: &schema.Resource{
-				Schema: equinix_schema.OperationalErrorSch(),
+				Schema: equinix_schema.ErrorSch(),
 			},
 		},
 	}
